pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"


[tool.poetry]
name = "python-base-graphql-api"
description = "Python Base GraphQL Project"
version = "0.1.0"
readme = "README.md"
authors = ["myapp.com"]


[tool.poetry.dependencies]
python = "^3.8.10"

django = "^3.2.3"
django-split-settings = "^1.0"
django-axes = "^5.16"
django-csp = "^3.7"
django-health-check = "^3.16"
django-http-referrer-policy = "^1.1"
django-permissions-policy = "^4.0"
django-stubs-ext = "^0.2"
django-storages = "^1.11.1"
django-filter = "^2.4.0"
django-heroku = "^0.3.1"
django-redis = "^5.0.0"

psycopg2-binary = "^2.8"
gunicorn = "^20.0"
python-decouple = "^3.4"
bcrypt = "^3.2"
structlog = "^21.1"
PyJWT = "^2.1.0"
Pillow = "^8.2.0"
boto3 = "^1.17.92"

graphene = {version = "^3.0.0b7", allow-prereleases = true}
graphene-django = {version = "3.0b7", allow-prereleases = true}
graphql-core = "^3.1.5"
celery = "^5.1.2"
hypothesis = "^6.14.5"
snapshottest = "^0.6.0"
django-celery-beat = "^2.2.1"



[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2"
django-querycount = "^0.7"
django-migration-linter = "^2.5"
django-extra-checks = "^0.11"
django-coverage-plugin = "^1.8"
nplusone = "^1.0"

wemake-python-styleguide = "^0.15"
flake8-pytest-style = "^1.4"
flake8-django = "^1.1"
flake8-logging-format = "^0.6"
autopep8 = "^1.5.7"
nitpick = "^0.26"

hypothesis = "^6.13"

django-stubs = "^1.8"

sphinx = "^4.0"
sphinx-autodoc-typehints = "^1.12"
tomlkit = "^0.7"
doc8 = "^0.8"

yamllint = "^1.26"
safety = "^1.10"
dotenv-linter = "^0.2"
dump-env = "^1.3"
ipython = "^7.23"
gevent = "^21.1.2"
mypy = "^0.941"
